add log output; add some tests

+++ examples/sample1/sample1.cpp
@@ -11,11 +11,11 @@ class Recevier : public SocketCanListener
 	void recvMessage( const CANMessage& message )
 	{
 		std::cout << std::hex << message.can_id << " ";
-		std::cout << std::dec << "[" << message.can_dlc << "] ";
+		std::cout << std::dec << "[" << static_cast<uint32_t>(message.can_dlc) << "] ";
 
 		for( int i = 0; (i < message.can_dlc) && (i < MAX_DATA_LENGTH) ; ++i )
 		{
-			std::cout << std::hex << message.data[i] << " ";
+			std::cout << std::hex << static_cast <uint32_t>(message.data[i]) << " ";
 		}
 		std::cout << std::dec << std::endl;
 	}
